core/custom-python-image/requirements.txt
# Generated by core/scripts/generate-requirements.py
# DO NOT EDIT MANUALLY - this file is generated from plugin dependencies
#
# To regenerate after modifying plugin dependencies:
#   python core/scripts/generate-requirements.py

PyYAML>=6.0,<7.0
opentelemetry-api>=1.30,<2.0
opentelemetry-exporter-otlp>=1.30,<2.0
opentelemetry-instrumentation-asgi>=0.51b0,<1.0
opentelemetry-instrumentation-fastapi>=0.51b0,<1.0
opentelemetry-instrumentation-httpx>=0.51b0,<1.0
opentelemetry-instrumentation-langchain>=0.51b0,<1.0
opentelemetry-instrumentation-openai-v2>=0.1.0
opentelemetry-instrumentation-requests>=0.51b0,<1.0
opentelemetry-sdk>=1.30,<2.0
